+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.0
 * Added a way to whisper currently previewed items to other people.
 * Fixed weirdness with weapon previews.
+* Added slash command to open the dressing room. Type /dressup or /dressingroom.
 
 ## 2.1.7
 * Fixed a bug in previous update.